[WIP]: Testing the performance
